+++ e2e/commands/deprecate.e2e.1.js
@@ -1,8 +1,9 @@
 import { expect } from 'chai';
 import path from 'path';
 import Helper from '../e2e-helper';
+import * as fixtures from '../fixtures/fixtures';
 
-describe('bit deprecate command', function () {
+describe('bit deprecate and undeprecate commands', function () {
   this.timeout(0);
   const helper = new Helper();
   after(() => {
@@ -13,7 +14,7 @@ describe('bit deprecate command', function () {
     before(() => {
       helper.initNewLocalScope();
     });
-    it('Should not deprecate component if bit.json is corrupted', () => {
+    it('should not deprecate component if bit.json is corrupted', () => {
       helper.corruptBitJson();
       try {
         helper.deprecateComponent('bar/foo2');
@@ -24,14 +25,16 @@ describe('bit deprecate command', function () {
       expect(output).to.include(`${path.join(helper.localScopePath, 'bit.json')}`);
     });
   });
-  describe('with local scope and tagged components', () => {
+  describe('deprecate local tagged component', () => {
     let output;
+    let scopeAfterDeprecation;
     before(() => {
       helper.setNewLocalAndRemoteScopes();
       helper.createComponentBarFoo();
       helper.addComponentBarFoo();
       helper.tagComponentBarFoo();
       output = helper.deprecateComponent('bar/foo');
+      scopeAfterDeprecation = helper.cloneLocalScope();
     });
     it('should show deprecated component', () => {
       expect(output).to.have.string('deprecated components: bar/foo');
@@ -46,6 +49,20 @@ describe('bit deprecate command', function () {
       output = helper.listRemoteScope(false);
       expect(output).to.contain.string('bar/foo');
       expect(output).to.contain.string('[Deprecated]');
+    });
+    describe('undeprecate local component', () => {
+      before(() => {
+        helper.getClonedLocalScope(scopeAfterDeprecation);
+        output = helper.undeprecateComponent('bar/foo');
+      });
+      it('should indicate the undeprecated components', () => {
+        expect(output).to.have.string('undeprecated components: bar/foo');
+      });
+      it('bit list should not show the component as deprecated', () => {
+        const listOutput = helper.listLocalScope();
+        expect(listOutput).to.contain.string('bar/foo');
+        expect(listOutput).to.not.contain.string('Deprecated');
+      });
     });
   });
   describe('with remote scope', () => {
@@ -70,16 +87,26 @@ describe('bit deprecate command', function () {
       const depOutput = helper.deprecateComponent(`${helper.remoteScope}/bar/foo111`, '-r');
       expect(depOutput).to.contain.string(`missing components: ${helper.remoteScope}/bar/foo`);
     });
+    describe('undeprecate remote component', () => {
+      before(() => {
+        output = helper.undeprecateComponent(`${helper.remoteScope}/bar/foo`, '-r');
+      });
+      it('should indicate the undeprecated components', () => {
+        expect(output).to.have.string(`undeprecated components: ${helper.remoteScope}/bar/foo`);
+      });
+      it('bit list should not show the component as deprecated', () => {
+        const listOutput = helper.listLocalScope();
+        expect(listOutput).to.contain.string('bar/foo');
+        expect(listOutput).to.not.contain.string('Deprecated');
+      });
+    });
   });
   describe('with remote scope with dependencies', () => {
     before(() => {
       helper.setNewLocalAndRemoteScopes();
-      const isTypeFixture = "module.exports = function isType() { return 'got is-type'; };";
-      helper.createFile('utils', 'is-type.js', isTypeFixture);
+      helper.createFile('utils', 'is-type.js', fixtures.isType);
       helper.addComponentUtilsIsType();
-      const isStringFixture =
-        "const isType = require('./is-type.js'); module.exports = function isString() { return isType() +  ' and got is-string'; };";
-      helper.createFile('utils', 'is-string.js', isStringFixture);
+      helper.createFile('utils', 'is-string.js', fixtures.isString);
       helper.addComponentUtilsIsString();
       helper.tagAllComponents();
       helper.exportAllComponents();